% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pref_tables.R
\name{check_pref_tables}
\alias{check_pref_tables}
\title{Check that a pair of preference tables are compatible
Tests if the tables are valid:
- same length keys between tables
- all values in one table are keys in the other
- length compatibility within tables confirmed by first 2 checks}
\usage{
check_pref_tables(male_pref, female_pref)
}
\arguments{
\item{male_pref}{Male preference table (list or dataframe)}

\item{female_pref}{Female preference table (list or dataframe)}
}
\value{
TRUE if tables are valid, FALSE otherwise
}
\description{
Check that a pair of preference tables are compatible
Tests if the tables are valid:
- same length keys between tables
- all values in one table are keys in the other
- length compatibility within tables confirmed by first 2 checks
}
\examples{
check_pref_tables(male1,female1)
check_pref_tables(male_prefs1,female_prefs3) #incompatible dimensions

}
