% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_stable_marriages}
\alias{find_stable_marriages}
\title{Implementation of the fundamental algorithm for the stable marriage problem}
\usage{
find_stable_marriages(male_prefs, female_prefs, U)
}
\arguments{
\item{male_prefs}{male preference table (list or dataframe)}

\item{female_prefs}{female preference table (list or dataframe)}

\item{U}{string name for undesirable male}
}
\value{
a stable matching with male names and female values (named character vector)
}
\description{
A function that uses a pair of preference tables and runs the Fundamental Algorithm to find a stable matching.
}
\examples{
find_stable_marriages(male_prefs3, female_prefs3, "U")

}
