% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-marriage-f2.R
\name{find_stable_marriages_safe}
\alias{find_stable_marriages_safe}
\title{Safe Implementation of the fundamental algorithm for the stable marriage problem}
\usage{
find_stable_marriages_safe(male_pref, female_pref, U)
}
\description{
A function that takes a pair of preference tables, checks that they are compatible,
and if they are, runs the Fundamental Algorithm to find a stable matching. See \code{\link{find_stable_marriages}}.
}
